Improve chances of stair-generation on special levels.
Special levels containing ordinary rooms (eg: oracle, some minetowns,
quest filler levels), use special room generation code.
This code has a number of checks, which if failed, cause room-generation
to abort.  Despite retrying 100 times for each room, it was common for
rooms not to be generated.  Sometimes the room that failed to generate
was the one containing the stairs.  This fixes one obvious reason for
the failure, and increases the retry count.  Room generation can still
fail, but usually for lack of space, after stairs are already done.