Fix: buffer not cleared for various functions that require player input.
SporkHack is old enough, EDIT_GETLINE does not exist, and typically
that's commented out as a compile time option. So by default, functions
like making a wish, or using #annotate, would remember the last input.
Some see this as a feature, but it's not common - if you're not
expecting this behavor and start typing, your last input is already
present in the buffer and you're adding to it, potentially sending
something the game doesn't understand (read: ruined wish).

The only edit that fixes this is in getline.c - the other file edits are
1) formatting bits I fixed while poking around in other files and 2)
setting makefiles and config.h for my local build. I'll revert all that
after I'm done working on 0.7.0 and am ready for it to be played
publically.